requirements.txt
Django 6.0.2
djangorestframework-3.16.1